Fix CI failures, undefined wolfSSL references, and add .gitignore
- Added .gitignore to exclude build artifacts and IDE metadata.
- Updated CI workflow to install dependencies (libuv1-dev, libssl-dev, libwolfssl-dev) on Ubuntu.
- Fixed Windows-specific function signatures and stdint.h conflicts.
- Resolved getentropy compilation error on Linux by using syscall(SYS_getrandom, ...).
- Fixed undefined references to HC128 and Rabbit in wolfSSL by adding feature guards.
- Fixed wolfSSL type mismatches (WC_RNG, Md5) and included missing headers.
- Resolved various format string warnings and parameter type inconsistencies.

+++ src/client.c
@@ -803,7 +803,7 @@ int do_kill(client_ctx *cx)
     conn *c;
 #if defined(NDEBUG)
 #else
-    pr_info("%s %d", __FUNCTION__, nread);
+    pr_info("%s %ld", __FUNCTION__, (long)nread);
 #endif
 
     c = CONTAINER_OF(handle, conn, handle);
